# OPENAPI_PAYMENTS_COMMISSIONS.yaml
openapi: 3.0.3
info:
  title: Changánet - Integrated Payments and Commissions API
  description: |
    Complete API documentation for the Integrated Payments and Commissions module of Changánet platform.

    ## Overview
    This API provides comprehensive payment processing, commission management, withdrawal handling, and financial audit capabilities for the Changánet professional services platform.

    ## Key Features
    - **Secure Payment Processing**: Integration with MercadoPago for payment processing with fund custody
    - **Commission Management**: Configurable commission rates (5-10%) applied only after service completion
    - **Professional Withdrawals**: Secure fund withdrawals to bank accounts with validation
    - **Financial Audit Trail**: Complete logging of all financial transactions and administrative actions
    - **Rate Limiting**: Configurable rate limits per endpoint type for security
    - **Webhook Security**: HMAC signature verification for payment notifications

    ## Authentication
    All endpoints require JWT Bearer token authentication except webhooks and auto-release endpoints.

    ## Rate Limiting
    - **Public endpoints**: 100 requests/15min (production), 1000 requests/15min (development)
    - **Authentication**: 5 requests/15min (production), 50 requests/15min (development)
    - **API endpoints**: 60 requests/min (production), 300 requests/min (development)
...
    - **Webhooks**: 20 requests/min (all environments)

    ## Error Codes
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  schemas:
...
    CreatePaymentPreferenceRequest:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
          description: ID of the service to pay for
          example: "550e8400-e29b-41d4-a716-446655440000"
        amount:
          type: number
          minimum: 500
          maximum: 500000
          description: Payment amount in ARS (minimum 500, maximum 500,000)
          example: 5000
        description:
          type: string
          maxLength: 255
          description: Payment description
          example: "Servicio de instalación eléctrica"
      required:
        - serviceId
        - amount
...
paths:
...
  /api/payments/create-preference:
    post:
      tags:
        - Payments
      summary: Create payment preference
      description: |
        Creates a MercadoPago payment preference for a service.

        **Requirements:**
        - User must be authenticated
        - User must be the client of the service
        - Service must not already have a payment
        - Amount must be between 500 and 500,000 ARS

        **Rate Limit:** 10 requests per 5 minutes (production)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentPreferenceRequest'
            examples:
              basic-payment:
                summary: Basic service payment
                value:
                  serviceId: "550e8400-e29b-41d4-a716-446655440000"
                  amount: 5000
                  description: "Servicio de instalación eléctrica"
      responses:
        '201':
          description: Payment preference created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentPreferenceResponse'
              examples:
                success:
                  summary: Successful preference creation
                  value:
                    success: true
                    data:
                      preferenceId: "550e8400-e29b-41d4-a716-446655440000"
                      paymentId: "550e8400-e29b-41d4-a716-446655440001"
                      initPoint: "https://www.mercadopago.com.ar/checkout/v1/redirect?pref_id=550e8400-e29b-41d4-a716-446655440000"
                      expirationDate: "2025-11-29T01:37:26.147Z"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid-service:
                  summary: Invalid service ID
                  value:
                    success: false
                    error: "Servicio no encontrado o no pertenece al usuario"
                    code: "INVALID_SERVICE"
                invalid-amount:
                  summary: Amount out of range
                  value:
                    success: false
                    error: "El monto debe estar entre 500 y 500,000 ARS"
                    code: "INVALID_AMOUNT"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no-auth:
                  summary: Missing authentication
                  value:
                    success: false
                    error: "Token de autenticación requerido"
                    code: "AUTH_REQUIRED"
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not-client:
                  summary: User is not the service client
                  value:
                    success: false
                    error: "No tienes permiso para pagar este servicio"
                    code: "INSUFFICIENT_PERMISSIONS"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                server-error:
                  summary: Unexpected server error
                  value:
                    success: false
                    error: "Error interno del servidor"
                    code: "INTERNAL_ERROR"
  /api/payments/status/{paymentId}:
    get:
      tags:
        - Payments
      summary: Get payment status
      description: |
        Retrieves the current status and details of a payment.

        **Requirements:**
        - User must be authenticated
        - User must be involved in the payment (client or professional)

        **Rate Limit:** 60 requests per minute (production)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatusResponse'
              examples:
                pending-payment:
                  summary: Payment in pending status
                  value:
                    success: true
                    data:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      serviceId: "550e8400-e29b-41d4-a716-446655440001"
                      status: "PENDIENTE"
                      amount: 5000
                      commission: 0
                      professionalAmount: 5000
                      createdAt: "2025-11-28T01:37:26.147Z"
                      updatedAt: "2025-11-28T01:37:26.147Z"
                approved-payment:
                  summary: Payment approved and ready for release
                  value:
                    success: true
                    data:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      serviceId: "550e8400-e29b-41d4-a716-446655440001"
                      status: "APROBADO"
                      amount: 5000
                      commission: 0
                      professionalAmount: 5000
                      createdAt: "2025-11-28T01:35:26.147Z"
                      updatedAt: "2025-11-28T01:37:26.147Z"
                released-payment:
                  summary: Payment released to professional
                  value:
                    success: true
                    data:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      serviceId: "550e8400-e29b-41d4-a716-446655440001"
                      status: "LIBERADO"
                      amount: 5000
                      commission: 250
                      professionalAmount: 4750
                      createdAt: "2025-11-28T01:35:26.147Z"
                      updatedAt: "2025-11-28T01:40:26.147Z"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not involved in payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/disputes:
    get:
      tags:
        - Payments
      summary: Get user disputes
      description: |
        Retrieves all payment disputes for the authenticated user.

        **Requirements:**
        - User must be authenticated

        **Rate Limit:** 60 requests per minute (production)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Disputes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                user-disputes:
                  summary: User's payment disputes
                  value:
                    success: true
                    data:
                      - id: "550e8400-e29b-41d4-a716-446655440000"
                        paymentId: "550e8400-e29b-41d4-a716-446655440001"
                        reason: "Servicio no completado"
                        status: "abierta"
                        createdAt: "2025-11-28T01:37:26.147Z"
                      - id: "550e8400-e29b-41d4-a716-446655440002"
                        paymentId: "550e8400-e29b-41d4-a716-446655440003"
                        reason: "Calidad del trabajo insuficiente"
                        status: "resuelta"
                        resolution: "reembolso_parcial"
                        createdAt: "2025-11-25T01:37:26.147Z"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/auto-release:
    post:
      tags:
        - Payments
      summary: Auto-release funds (system)
      description: |
        Automatically releases funds for services completed more than 24 hours ago.

        **Security:**
        - No authentication required (internal system operation)
        - Should be called by cron jobs only

        **Process:**
        - Finds all approved payments for completed services older than 24 hours
        - Applies commission and releases funds
        - Updates payment status to LIBERADO

        **Rate Limit:** Not applicable (system operation)
      responses:
        '200':
          description: Auto-release completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                auto-release-success:
                  summary: Funds auto-released successfully
                  value:
                    success: true
                    data:
                      processedPayments: 5
                      totalCommission: 1250
                      totalReleased: 23750
                      processedAt: "2025-11-28T01:37:26.147Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Commission Management Endpoints
...
  /api/payments/release-funds:
    post:
      tags:
        - Payments
      summary: Release funds to professional
      description: |
        Releases payment funds to the professional after service completion.

        **Requirements:**
        - User must be authenticated as the service client
        - Service must be marked as completed
        - Payment must be in approved status
        - Funds are held in custody until this point

        **Commission Application:**
        - Platform commission (5-10%) is applied only when funds are released
        - Commission is calculated on the total payment amount
        - Professional receives: total_amount - commission

        **Rate Limit:** 10 requests per 5 minutes (production)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseFundsRequest'
            examples:
              release-funds:
                summary: Release funds for completed service
                value:
                  paymentId: "123456789"
                  serviceId: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Funds released successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseFundsResponse'
              examples:
                success:
                  summary: Funds released with commission applied
                  value:
                    success: true
                    data:
                      commission: 250
                      professionalAmount: 4750
                      commissionRate: 0.05
                      paymentId: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          description: Invalid request or business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                service-not-completed:
                  summary: Service not marked as completed
                  value:
                    success: false
                    error: "El servicio debe estar completado para liberar fondos"
                    code: "SERVICE_NOT_COMPLETED"
                invalid-payment:
                  summary: Payment not in valid state
                  value:
                    success: false
                    error: "El pago no está en estado válido para liberación"
                    code: "INVALID_PAYMENT_STATUS"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the service client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not-client:
                  summary: User is not the service client
                  value:
                    success: false
                    error: "Solo el cliente puede liberar los fondos"
                    code: "INSUFFICIENT_PERMISSIONS"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/webhook:
    post:
      tags:
        - Payments
      summary: MercadoPago webhook handler
      description: |
        Handles payment status updates from MercadoPago.

        **Security:**
        - HMAC signature verification required
        - Timestamp validation (max 5 minutes old)
        - Rate limiting applied

        **Supported Events:**
        - `approved`: Payment approved, funds in custody
        - `rejected`: Payment rejected by gateway
        - `cancelled`: Payment cancelled by user
        - `expired`: Payment preference expired

        **Rate Limit:** 20 requests per minute (all environments)
      parameters:
        - name: x-signature
          in: header
          required: true
          schema:
            type: string
          description: MercadoPago HMAC signature (format ts=123456789,v1=signature)
          example: "ts=1732757046,v1=abc123def456"
        - name: x-request-id
          in: header
          required: true
          schema:
            type: string
          description: Unique request identifier
          example: "req_550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
            examples:
              payment-approved:
                summary: Payment approved notification
                value:
                  id: 123456789
                  status: "approved"
                  status_detail: "accredited"
                  external_reference: "550e8400-e29b-41d4-a716-446655440000"
                  transaction_amount: 5000
                  payment_type_id: "credit_card"
                  date_approved: "2025-11-28T01:37:26.147Z"
                  date_created: "2025-11-28T01:35:26.147Z"
              payment-rejected:
                summary: Payment rejected notification
                value:
                  id: 123456789
                  status: "rejected"
                  status_detail: "cc_rejected_bad_filled_other"
                  external_reference: "550e8400-e29b-41d4-a716-446655440000"
                  transaction_amount: 5000
                  date_created: "2025-11-28T01:35:26.147Z"
      responses:
        '200':
          description: Webhook processed successfully
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        '400':
          description: Invalid webhook data or signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid-signature:
                  summary: Invalid HMAC signature
                  value:
                    success: false
                    error: "Invalid webhook signature"
                    code: "INVALID_SIGNATURE"
                timestamp-too-old:
                  summary: Webhook timestamp too old
                  value:
                    success: false
                    error: "Webhook timestamp too old"
                    code: "TIMESTAMP_TOO_OLD"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/withdraw:
    post:
      tags:
        - Payments
      summary: Request fund withdrawal
      description: |
        Allows professionals to withdraw available funds to their bank account.

        **Requirements:**
        - User must be authenticated as a professional
        - Must have available funds from completed services
        - Bank details must be valid (CVU format, alias)
        - Amount must be between 100 and 50,000 ARS

        **Process:**
        - Withdrawal request is created and queued for processing
        - Bank transfer is processed within 2-3 business days
        - Professional receives notification via email/app

        **Rate Limit:** 10 requests per 5 minutes (production)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawFundsRequest'
            examples:
              withdrawal-request:
                summary: Professional withdrawal request
                value:
                  amount: 5000
                  bankDetails:
                    cvu: "1234567890123456789012"
                    alias: "mi.cuenta.bancaria"
      responses:
        '200':
          description: Withdrawal request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawFundsResponse'
              examples:
                success:
                  summary: Withdrawal request accepted
                  value:
                    success: true
                    data:
                      withdrawalId: "550e8400-e29b-41d4-a716-446655440000"
                      amount: 5000
                      processedAt: "2025-11-28T01:37:26.147Z"
                      estimatedArrival: "2025-12-02T01:37:26.147Z"
                      bankDetails:
                        alias: "mi.cuenta.bancaria"
                        cvuMasked: "***56789012"
                      status: "PROCESANDO"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid-cvu:
                  summary: Invalid CVU format
                  value:
                    success: false
                    error: "El CVU debe tener exactamente 22 dígitos"
                    code: "INVALID_CVU"
                invalid-amount:
                  summary: Amount out of range
                  value:
                    success: false
                    error: "El monto debe estar entre 100 y 50,000 ARS"
                    code: "INVALID_AMOUNT"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a professional
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not-professional:
                  summary: User is not a professional
                  value:
                    success: false
                    error: "Solo los profesionales pueden retirar fondos"
                    code: "INSUFFICIENT_PERMISSIONS"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient-funds:
                  summary: Not enough available funds
                  value:
                    success: false
                    error: "Fondos insuficientes para el retiro solicitado"
                    code: "INSUFFICIENT_FUNDS"
  /api/payments/receipt/{paymentId}:
    get:
      tags:
        - Payments
      summary: Generate payment receipt
      description: |
        Generates and returns a payment receipt for completed transactions.

        **Requirements:**
        - User must be authenticated
        - User must be involved in the payment
        - Payment must be completed (approved or released)

        **Rate Limit:** 60 requests per minute (production)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Receipt generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
              examples:
                receipt-pdf:
                  summary: PDF receipt file
                  value: "%PDF-1.4..."
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not involved in payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payment not found or no receipt available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/receipts/{fileName}:
    get:
      tags:
        - Payments
      summary: Download payment receipt
      description: |
        Downloads a previously generated payment receipt file.

        **Security:**
        - Filename validation to prevent path traversal
        - User must be authenticated

        **Rate Limit:** 60 requests per minute (production)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileName'
      responses:
        '200':
          description: Receipt file downloaded successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/{paymentId}/dispute:
    post:
      tags:
        - Payments
      summary: Create payment dispute
      description: |
        Creates a dispute for a payment that requires resolution.

        **Requirements:**
        - User must be authenticated
        - User must be involved in the payment
        - Payment must be in a disputable state
        - Maximum 5 evidence files allowed

        **Rate Limit:** 10 requests per 5 minutes (production)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisputeRequest'
            examples:
              service-dispute:
                summary: Dispute for incomplete service
                value:
                  reason: "Servicio no completado"
                  description: "El profesional no realizó el trabajo acordado"
                  evidence:
                    - "https://storage.googleapis.com/evidence/photo1.jpg"
                    - "https://storage.googleapis.com/evidence/contract.pdf"
      responses:
        '201':
          description: Dispute created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                dispute-created:
                  summary: Dispute successfully created
                  value:
                    success: true
                    data:
                      disputeId: "550e8400-e29b-41d4-a716-446655440000"
                      status: "abierta"
                      createdAt: "2025-11-28T01:37:26.147Z"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not involved in payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...